+++ src/main/java/com/networknt/schema/MaximumValidator.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2016 Network New Technologies Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
- * You may not use this file except in compliance with the License.
+ * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
@@ -20,6 +20,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.Set;
 
@@ -27,15 +28,15 @@ public class MaximumValidator extends BaseJsonValidator implements JsonValidator
     private static final Logger logger = LoggerFactory.getLogger(MaximumValidator.class);
     private static final String PROPERTY_EXCLUSIVE_MAXIMUM = "exclusiveMaximum";
 
-    private double maximum;
     private boolean excludeEqual = false;
 
-    public MaximumValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
+    private final ThresholdMixin typedMaximum;
 
+
+    public MaximumValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
         super(schemaPath, schemaNode, parentSchema, ValidatorTypeCode.MAXIMUM, validationContext);
-        if (schemaNode.isNumber()) {
-            maximum = schemaNode.doubleValue();
-        } else {
+
+        if (!schemaNode.isNumber()) {
             throw new JsonSchemaException("maximum value is not a number");
         }
 
@@ -45,24 +46,68 @@ public class MaximumValidator extends BaseJsonValidator implements JsonValidator
         }
 
         parseErrorCode(getValidatorType().getErrorCodeKey());
+
+        if (!JsonType.INTEGER.toString().equals(getNodeFieldType())) {
+            // "number" or no type
+            // by default treat value as double: compatible with previous behavior
+            final double dm = schemaNode.doubleValue();
+            typedMaximum = new ThresholdMixin() {
+                @Override
+                public boolean crossesThreshold(JsonNode node) {
+                    double value = node.asDouble();
+                    return greaterThan(value, dm) || (excludeEqual && MaximumValidator.this.equals(value, dm));
+                }
+
+                @Override
+                public String thresholdValue() {
+                    return String.valueOf(dm);
+                }
+            };
+
+        } else if ( schemaNode.isLong() || schemaNode.isInt() ) {
+            // "integer", and within long range
+            final long lm = schemaNode.asLong();
+            typedMaximum = new ThresholdMixin() {
+                @Override
+                public boolean crossesThreshold(JsonNode node) {
+                    long val = node.asLong();
+                    return node.isBigInteger() || lm < val || (excludeEqual && lm <= val);
+                }
+
+                @Override
+                public String thresholdValue() {
+                    return String.valueOf(lm);
+                }
+            };
+
+        } else {
+            // "integer" outside long range
+            final BigInteger bim = new BigInteger(schemaNode.asText());
+            typedMaximum = new ThresholdMixin() {
+                @Override
+                public boolean crossesThreshold(JsonNode node) {
+                    int cmp = bim.compareTo(node.bigIntegerValue());
+                    return cmp < 0 || (excludeEqual && cmp <= 0);
+                }
+
+                @Override
+                public String thresholdValue() {
+                    return String.valueOf(bim);
+                }
+            };
+        }
     }
 
     public Set<ValidationMessage> validate(JsonNode node, JsonNode rootNode, String at) {
         debug(logger, node, rootNode, at);
 
-        if (!node.isNumber()) {
+        if (!TypeValidator.isNumber(node, config.isTypeLoose())) {
             // maximum only applies to numbers
             return Collections.emptySet();
         }
 
-        String fieldType = this.getNodeFieldType();
-
-        double value = node.doubleValue();
-        if (greaterThan(value, maximum) || (excludeEqual && equals(value, maximum))) {
-            if (JsonType.INTEGER.toString().equals(fieldType)) {
-                return Collections.singleton(buildValidationMessage(at, "" + (int)maximum));
-            }
-            return Collections.singleton(buildValidationMessage(at, "" + maximum));
+        if (typedMaximum.crossesThreshold(node)) {
+            return Collections.singleton(buildValidationMessage(at, typedMaximum.thresholdValue()));
         }
         return Collections.emptySet();
     }