+++ src/main/java/com/networknt/schema/ValidationContext.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2016 Network New Technologies Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.networknt.schema;
 
 import com.fasterxml.jackson.databind.JsonNode;
@@ -5,6 +21,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 public class ValidationContext {
     private final JsonMetaSchema metaSchema;
     private final JsonSchemaFactory jsonSchemaFactory;
+    private SchemaValidatorsConfig config;
     
     public ValidationContext(JsonMetaSchema metaSchema, JsonSchemaFactory jsonSchemaFactory) {
         if (metaSchema == null) {
@@ -18,11 +35,19 @@ public class ValidationContext {
     }
     
     public JsonValidator newValidator(String schemaPath, String keyword /* keyword */, JsonNode schemaNode,
-            JsonSchema parentSchema) {
+                                      JsonSchema parentSchema) {
         return metaSchema.newValidator(this, schemaPath, keyword, schemaNode, parentSchema);
     }
     
     public JsonSchemaFactory getJsonSchemaFactory() {
         return jsonSchemaFactory;
+    }
+
+    public SchemaValidatorsConfig getConfig() {
+        return config;
+    }
+
+    public void setConfig(SchemaValidatorsConfig config) {
+        this.config = config;
     }
 }